# backend/api-contract.yaml
openapi: 3.1.0
info:
  title: SIGAP API
  description: >
    API Contract untuk SIGAP (Sistem Integrasi Gizi & Akuntabilitas Pangan) —
    platform tata kelola vendor Program Makan Bergizi Gratis (MBG). Mencakup 3
    modul inti fase MVP: Portal Perizinan Vendor, Dashboard Monitoring
    Distribusi, dan Sistem Pengaduan Publik.


    Versi ini (0.2.0) memperbaiki sejumlah isu dari draft 0.1.0: schema yang
    salah tempat di bawah components.responses, duplicate key YAML yang
    menghilangkan schema DistributionReport, kebocoran NIK mentah lewat
    schema Vendor, radius geospasial yang bisa dimanipulasi lewat payload
    client, serta idempotency key yang tidak konsisten diterapkan di seluruh
    endpoint state-changing.
  version: 0.2.0
  contact:
    name: Tim SIGAP
servers:
  - url: https://api.sigap.example.id/v1
    description: Production (placeholder, sesuaikan dengan domain final)
  - url: http://localhost:8000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /vendors:
    post:
      tags:
        - Vendor
      summary: Registrasi vendor baru
      description: >
        Endpoint ini menerima data registrasi vendor. `nik_penanggung_jawab`
        yang dikirim di request disimpan terenkripsi (pgcrypto AES-256) dan
        TIDAK PERNAH dikembalikan mentah lewat response ini maupun endpoint
        vendor lain — response selalu memakai schema `Vendor` yang hanya
        berisi `nik_penanggung_jawab_masked`. Untuk kebutuhan reveal NIK
        mentah oleh admin/verifikator, gunakan endpoint khusus
        GET /vendors/{vendorId}/nik yang wajib tercatat di audit log.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorRegistrationRequest'
      responses:
        '201':
          description: Vendor tersimpan dengan status pending_verification
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      tags:
        - Vendor
      summary: Daftar vendor dengan filter scope
      description: >
        Hasil otomatis ter-scope sesuai role (mis. Pengawas Dinas Jakarta
        hanya melihat vendor di wilayah Jakarta). Parameter `province` dapat
        ditambahkan untuk memperjelas filter wilayah bila diperlukan. Endpoint
        baca (GET) tidak memerlukan X-Idempotency-Key — idempotency hanya
        relevan untuk operasi yang mengubah state.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending_verification
              - verified
              - rejected
              - suspended
        - name: province
          in: query
          description: Filter vendor berdasarkan provinsi (ABAC)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vendor'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /vendors/{vendorId}:
    parameters:
      - name: vendorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Vendor
      summary: Detail satu vendor
      description: >
        Response selalu memakai schema `Vendor` (NIK ter-mask). Ini berlaku
        untuk semua role, termasuk admin/verifikator — lihat
        GET /vendors/{vendorId}/nik untuk reveal NIK mentah secara terpisah.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Vendor
      summary: Update data vendor
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
...
  /vendors/{vendorId}/verify:
    post:
      tags:
        - Vendor
      summary: >
        Verifikasi & keputusan approval vendor oleh Verifikator BGN. Approve
        memicu penerbitan SIO Digital.
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  enum:
                    - approve
                    - reject
                notes:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /distributions:
    post:
      tags:
        - Distribution
      summary: Submit laporan distribusi harian, wajib disertai foto & koordinat GPS
      description: >
        Client HANYA mengirim latitude/longitude/foto/metadata EXIF. Radius
        jangkauan sekolah, toleransi, dan hasil validasi geospasial
        SEPENUHNYA dihitung di server (ST_DWithin, DATA_GOVERNANCE.md §5)
        berdasarkan data sekolah yang terdaftar — client tidak dapat
        mengirim/mengubah nilai radius lewat payload ini.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DistributionReportRequest'
      responses:
        '201':
          description: Laporan tersimpan; anomaly detection berjalan asinkron
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionReport'
    get:
      tags:
        - Distribution
      summary: Daftar laporan distribusi, ter-scope sesuai role
      parameters:
        - name: vendor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: has_anomaly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DistributionReport'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /distributions/{id}:
    get:
      tags:
        - Distribution
      summary: Detail satu laporan distribusi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionReport'
  /distributions/{id}/metadata:
    get:
      tags:
        - Distribution
      summary: Metadata teknis lengkap laporan distribusi untuk keperluan audit
      description: >
        Menyediakan EXIF, capture time, dan lokasi lengkap bagi auditor untuk
        investigasi dugaan tampering foto/lokasi (DATA_GOVERNANCE.md §6).
        Dibatasi role verifikator_bgn/pengawas_dinas/admin.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionMetadata'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /distributions/{id}/appeal:
    post:
      tags:
        - Distribution
      summary: Vendor mengajukan banding atas flag anomali yang dihasilkan AI
      description: >
        Selama status banding `pending_review`/`in_progress`, perhitungan
        Vendor Score untuk laporan ini dibekukan (`is_frozen=true` pada
        object anomaly) sampai banding diputuskan.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                supporting_evidence:
                  type: string
                  format: binary
      responses:
        '200':
          description: Banding diajukan, status vendor score dibekukan sementara
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppealResponse'
  /complaints:
    post:
      tags:
        - Complaint
      summary: Submit pengaduan masyarakat. Bisa tanpa login (pelapor dapat anonim).
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ComplaintRequest'
      responses:
        '201':
          description: Pengaduan tersimpan, mendapat nomor tiket untuk tracking
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
    get:
      tags:
        - Complaint
      summary: Daftar pengaduan untuk verifikator/pengawas, ter-scope wilayah
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - baru
              - diproses
              - ditindaklanjuti
              - ditutup
        - name: severity
          in: query
          schema:
            type: string
            enum:
              - rendah
              - sedang
              - tinggi
              - kritis
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Complaint'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /complaints/{id}:
    get:
      tags:
        - Complaint
      summary: >
        Detail pengaduan & status tracking. Nomor tiket berfungsi sebagai
        kunci akses publik untuk melacak laporan sendiri.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
    patch:
      tags:
        - Complaint
      summary: Update status & tindak lanjut pengaduan (khusus verifikator BGN)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - diproses
                    - ditindaklanjuti
                    - ditutup
                resolution_notes:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
...
components:
  parameters:
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: >
        WAJIB (required, bukan opsional) untuk semua operasi yang mengubah
        state (POST/PATCH/DELETE) guna mencegah duplicate submission akibat
        retry/network lag/klik ganda (DATA_GOVERNANCE.md §11). Tidak dipakai
        pada operasi baca (GET). Server menolak request tanpa header ini,
        atau dengan key yang sama dalam window 24 jam (kecuali sebagai
        retry dari request identik).
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        example: 550e8400-e29b-41d4-a716-446655440000
  headers:
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
    RateLimitRemaining:
      description: Remaining requests in the current window
      schema:
        type: integer
    RateLimitReset:
      description: Timestamp (epoch seconds) when the rate-limit window resets
      schema:
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request tidak valid (validasi payload gagal)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Role/scope pengguna tidak memiliki izin atas resource ini
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: >
        Rate limit exceeded. Client should respect the RateLimit-Reset
        header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: true
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...
    VendorRegistrationRequest:
      type: object
      required:
        - nama_usaha
        - nik_penanggung_jawab
        - nib
        - alamat
        - provinsi
        - kabupaten_kota
      properties:
        nama_usaha:
          type: string
        nik_penanggung_jawab:
          type: string
          example: 3175xxxxxxxxxxxx
          description: >
            NIK mentah, hanya dipakai sekali saat registrasi untuk disimpan
            terenkripsi (pgcrypto AES-256). Tidak pernah dikembalikan mentah
            oleh endpoint manapun kecuali GET /vendors/{vendorId}/nik.
        nib:
          type: string
        alamat:
          type: string
        provinsi:
          type: string
        kabupaten_kota:
          type: string
        kontak_telepon:
          type: string
    VendorUpdateRequest:
      type: object
      description: >
        Sengaja tidak menerima field NIK — perubahan data identitas legal
        harus lewat proses verifikasi ulang, bukan update biasa.
      properties:
        alamat:
          type: string
        kontak_telepon:
          type: string
    Vendor:
      type: object
      description: >
        Schema standar untuk SEMUA role (termasuk admin/verifikator). NIK
        mentah tidak pernah muncul sebagai field di schema ini — hanya versi
        masked. Ini bukan konvensi penyembunyian di frontend, tapi memang
        tidak ada field NIK mentah di object ini sama sekali, sehingga tidak
        bisa bocor lewat inspect response/DevTools.
      properties:
        id:
          type: string
          format: uuid
        nama_usaha:
          type: string
        nik_penanggung_jawab_masked:
          type: string
          example: 3175********1234
          description: >
            NIK ter-mask (4 digit awal + 4 digit akhir, sisanya bintang).
            Satu-satunya representasi NIK yang beredar lewat endpoint umum.
        nib:
          type: string
        alamat:
          type: string
        provinsi:
          type: string
        kabupaten_kota:
          type: string
        status:
          type: string
          enum:
            - pending_verification
            - verified
            - rejected
            - suspended
        vendor_score:
          type: number
          format: float
          example: 87.5
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    DistributionReportRequest:
      type: object
      description: >
        Client HANYA mengirim latitude/longitude/foto/metadata. Field radius
        SENGAJA tidak ada di request ini — radius jangkauan sekolah dan
        toleransi ditentukan & divalidasi sepenuhnya di server berdasarkan
        data sekolah yang terdaftar, agar tidak bisa dimanipulasi lewat
        payload (mis. mengecilkan/membesarkan radius supaya laporan palsu
        lolos validasi).
      required:
        - vendor_id
        - jumlah_porsi
        - latitude
        - longitude
        - foto
      properties:
        vendor_id:
          type: string
          format: uuid
        jumlah_porsi:
          type: integer
        lokasi_sekolah:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        foto:
          type: string
          format: binary
        metadata:
          type: object
          description: Metadata waktu pengambilan foto & device info
          required:
            - capture_time
          properties:
            capture_time:
              type: string
              format: date-time
              description: Timestamp aktual pengambilan foto (EXIF).
            exif_timestamp:
              type: string
              format: date-time
              nullable: true
            device_id:
              type: string
              nullable: true
    DistributionReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vendor_id:
          type: string
          format: uuid
        jumlah_porsi:
          type: integer
        lokasi_sekolah:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        foto_url:
          type: string
          format: uri
        reported_at:
          type: string
          format: date-time
          description: >
            Waktu server menerima laporan. Bukan waktu foto diambil — lihat
            photo_taken_at untuk itu.
        photo_taken_at:
          type: string
          format: date-time
          nullable: true
          description: >
            Timestamp EXIF asli saat foto diambil. Jika selisih dengan
            created_at lebih dari 24 jam, tampering_suspicion otomatis true
            (DATA_GOVERNANCE.md §6) — mencegah kasus foto lama dikirim ulang
            seolah-olah baru.
        tampering_suspicion:
          type: boolean
          description: >
            True jika photo_taken_at lebih lama dari created_at >24 jam,
            atau metadata EXIF hilang/tidak konsisten.
        geo_validation:
          type: object
          description: >
            Hasil validasi radius geospasial, dihitung sepenuhnya di server
            (ST_DWithin) — tidak bergantung pada input client sama sekali.
          properties:
            distance_from_school_meters:
              type: number
              format: float
            radius_allowed_meters:
              type: number
              format: float
              description: Radius jangkauan resmi sekolah, dikonfigurasi di server.
            radius_tolerance_meters:
              type: number
              format: float
              example: 50
              description: Toleransi default (DATA_GOVERNANCE.md §5).
            within_tolerance:
              type: boolean
            requires_manual_verification:
              type: boolean
              description: >
                True jika jarak di luar toleransi ±10%, ditandai untuk
                verifikasi manual.
        anomaly:
          type: object
          nullable: true
          description: Hasil analisis dari Anomaly Detection Engine (Probabilistik)
          properties:
            score:
              type: number
              format: float
              description: Probabilitas anomali 0.0 - 1.0
            confidence:
              type: number
              format: float
            flag:
              type: string
              enum:
                - none
                - low_risk
                - medium_risk
                - high_risk
                - critical
            detected:
              type: boolean
              description: True jika flag berstatus high_risk atau critical
            details:
              type: array
              items:
                type: string
              description: Faktor-faktor yang berkontribusi pada skor
            appeal_status:
              type: string
              enum:
                - none
                - pending_review
                - in_progress
                - resolved
                - rejected
              default: none
            is_frozen:
              type: boolean
              description: >
                True jika perhitungan vendor score dibekukan karena banding
                sedang pending_review/in_progress.
    DistributionMetadata:
      type: object
      description: Detail teknis untuk audit dugaan tampering (DATA_GOVERNANCE.md §6).
      properties:
        distribution_id:
          type: string
          format: uuid
        photo_taken_at:
          type: string
          format: date-time
          nullable: true
        exif_timestamp:
          type: string
          format: date-time
          nullable: true
        device_id:
          type: string
          nullable: true
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        tampering_suspicion:
          type: boolean
        created_at:
          type: string
          format: date-time
    AppealResponse:
      type: object
      properties:
        distribution_id:
          type: string
          format: uuid
        appeal_status:
          type: string
          enum:
            - pending_review
            - in_progress
        is_frozen:
          type: boolean
          example: true
        submitted_at:
          type: string
          format: date-time
    ComplaintRequest:
      type: object
      required:
        - vendor_id
        - kategori
        - deskripsi
      properties:
        vendor_id:
          type: string
          format: uuid
        nama_pelapor:
          type: string
          description: Opsional — pelapor dapat mengirim tanpa identitas
        kategori:
          type: string
          enum:
            - keracunan
            - keterlambatan
            - kekurangan_porsi
            - kualitas_makanan
            - lainnya
        deskripsi:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        foto:
          type: string
          format: binary
        distribution_id:
          type: string
          format: uuid
          description: Opsional — ID laporan distribusi spesifik yang diadukan
        tanggal_kejadian:
          type: string
          format: date
          description: Tanggal kejadian (opsional, jika spesifik per hari)
    Complaint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_number:
          type: string
          example: SG-2026-000456
        vendor_id:
          type: string
          format: uuid
        kategori:
          type: string
        deskripsi:
          type: string
        severity:
          type: string
          enum:
            - rendah
            - sedang
            - tinggi
            - kritis
        status:
          type: string
          enum:
            - baru
            - diproses
            - ditindaklanjuti
            - ditutup
        resolution_notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        sla_deadline:
          type: string
          format: date-time
...